+++ src/app/components/UserInfoMenu/index.tsx
@@ -11,7 +11,7 @@ import * as React from 'react';
 import {IUser} from '../../repository/user/interface';
 import {CloseRounded, CheckRounded, EditRounded, AddRounded, KeyboardBackspaceRounded} from '@material-ui/icons';
 import IconButton from '@material-ui/core/IconButton/IconButton';
-import {InputPeer, PeerNotifySettings, PhoneContact} from '../../services/sdk/messages/chat.core.types_pb';
+import {InputPeer, PeerNotifySettings, PeerType, PhoneContact} from '../../services/sdk/messages/chat.core.types_pb';
 import SDK from '../../services/sdk';
 import UserAvatar from '../UserAvatar';
 import TextField from '@material-ui/core/TextField/TextField';
@@ -472,13 +472,21 @@ class UserInfoMenu extends React.Component<IProps, IState> {
         if (!user || !user.photo) {
             return;
         }
+        let peer: InputPeer | undefined;
+        if (user.accesshash) {
+            peer = new InputPeer();
+            peer.setAccesshash(user.accesshash);
+            peer.setId(user.id || '');
+            peer.setType(PeerType.PEERUSER);
+        }
         const doc: IDocument = {
             items: [{
                 caption: '',
                 fileLocation: user.photo.photobig,
                 thumbFileLocation: user.photo.photosmall,
             }],
-            type: 'avatar'
+            peer,
+            type: 'avatar',
         };
         this.documentViewerService.loadDocument(doc);
     }